Fix tag serial interface and add frame transforms
This gets the node much closer to a final version. Documentation and style changes also updated.

+++ src/decawave/include/decawave/decawave.h
@@ -12,9 +12,9 @@
 namespace decawave{
 struct Anchor{
   int id;
-  int distance;
+  double distance;
   int distance_quality;
-  int position[3];
+  double position[3];
   int quality_factor;
 };
